+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Initial release of Flutter Pytorch Plugin by A.F.K.M.
+
+## 1.0.0
+
+* Fixed the late initializer issue of Flutter Pytorch Plugin by A.F.K.M.